Reorganized aliases, added functions, changed path
I reorganized some of the aliases to reflect changes in how I do things
and I added a few more, particularly, a few for git (like gs for git
status) that came in handy when using it more frequently for a recent
project.

I also added a few convenience functions:

	hr n -> echo a line of n dashes.

	list-ips -> list the IPs of all connected network interfaces.

	pyserv -> start the Python web server module in current
	directory and print ips with list-ips.

	search-files-from [dir] [grep options] -> search all files,
	recursively, from the given "dir" with the given grep options
	(useful to put in particular flags like "-i" followed by the
	pattern to search for).

	search-files [grep options] -> Same as above, but search from
	the current directory.

Lastly, I changed how I update my path. Instead of relying on
.profile, since that requires a restart for changes to apply, and
making a change in both is stupid, I'm just listing updates to the path
in .bashrc from least-important to most-important. I left one path
update in .profile for reasons of using the .profile on other computers.