Updated bundles, variable names, added autogroups
I updated some bundle documents and added a few new ones, including a
few for LISP/Hy and Haskell.

I updated some of the variable names for the custom build function I
use, as well.

I added autogroups for Hy and Haskell and I created a few more commands
to be used with the custom build function as well as 2 key mappings for
toggling between two of the commands.

There are also a few small changes not worth mentioning in detail